feat(graph): enhance caching mechanism for notes annual and evaluation graphs
